Added authentication, moved pages around to be in order of workflow
