--- data/AMS/README.txt
@@ -1,18 +1,0 @@
-Documents are not uploaded to git. The list of documents which were uploaded to pinecone database AMS:
-AMS_1996, https://ntrs.nasa.gov/citations/19960025595
-AMS_1997, https://ntrs.nasa.gov/citations/19970021613
-AMS_1998, https://ntrs.nasa.gov/citations/19980193156
-AMS_1999, https://ntrs.nasa.gov/citations/19990053852
-AMS_2000, https://ntrs.nasa.gov/citations/20000048380
-AMS_2001, https://ntrs.nasa.gov/citations/20010071164
-AMS_2002, https://ntrs.nasa.gov/citations/20020050182
-AMS_2004, https://ntrs.nasa.gov/citations/20040084272
-AMS_2006, https://ntrs.nasa.gov/citations/20060028221
-AMS_2008, https://ntrs.nasa.gov/citations/20080023060
-AMS_2010, https://ntrs.nasa.gov/citations/20100021914
-AMS_2012, https://ntrs.nasa.gov/citations/20130008824
-AMS_2014, https://ntrs.nasa.gov/citations/20140008875
-AMS_2016, https://ntrs.nasa.gov/citations/20160004038
-AMS_2018, https://ntrs.nasa.gov/citations/20180002828
-AMS_2020, https://ntrs.nasa.gov/citations/20205009766
-AMS_2022, https://ntrs.nasa.gov/citations/20220006415
