+++ src/aerospace_chatbot/admin.py
@@ -1,16 +1,15 @@
+import data_processing
+from prompts import TEST_QUERY_PROMPT
+
 import os
 import logging
 import json
+import streamlit as st
 
 import openai
 from pinecone import Pinecone
 import chromadb
-
-import streamlit as st
-
 from langchain_openai import ChatOpenAI
-
-from prompts import TEST_QUERY_PROMPT
 
 # Set up the page, enable logging 
 from dotenv import load_dotenv,find_dotenv
@@ -418,7 +417,69 @@ def show_ragatouille_indexes(format=True):
         return _format_ragatouille_status(indexes)
     else:
         return indexes
-    
+
+def st_connection_status_expander(expanded: bool = True, delete_buttons: bool = False):
+    """
+    Expands a Streamlit expander widget to display connection status information.
+
+    Args:
+        expanded (bool, optional): Whether the expander is initially expanded or collapsed. Only intended with account access. Defaults to True.
+        delete_buttons (bool, optional): Whether to display delete buttons for Pinecone and Chroma DB indexes. Defaults to False.
+    """
+
+    with st.expander("Connection Status", expanded=expanded):
+        # Show key status
+        st.markdown("**API keys** (Indicates status of local variable. It does not guarantee the key itself is correct):")
+        st.markdown(test_key_status())
+
+        # Pinecone
+        st.markdown(show_pinecone_indexes())
+        try:
+            pinecone_indexes = [obj.name for obj in show_pinecone_indexes(format=False)['message']]
+            pinecone_index_name = st.selectbox('Pinecone index to delete', pinecone_indexes)
+            if delete_buttons:
+                if st.button('Delete Pinecone index', help='This is permanent!'):
+                    if pinecone_index_name.endswith("parent-child"):
+                        rag_type = "Parent-Child"
+                    elif pinecone_index_name.endswith("summary"):
+                        rag_type = "Summary"
+                    else:
+                        rag_type = "Standard"
+                    data_processing.delete_index('Pinecone', pinecone_index_name, rag_type, local_db_path='../../db')
+                    st.markdown(f"Index {pinecone_index_name} has been deleted.")
+        except:
+            pass
+
+        # Chroma DB
+        st.markdown(show_chroma_collections())
+        try:
+            chroma_db_collections = [obj.name for obj in show_chroma_collections(format=False)['message']]
+            if delete_buttons:
+                chroma_db_name = st.selectbox('Chroma database to delete', chroma_db_collections)
+                if st.button('Delete Chroma database', help='This is permanent!'):
+                    if chroma_db_name.endswith("parent-child"):
+                        rag_type = "Parent-Child"
+                    elif "-summary-" in chroma_db_name:
+                        rag_type = "Summary"
+                    else:
+                        rag_type = "Standard"
+                    data_processing.delete_index('ChromaDB', chroma_db_name, rag_type, local_db_path='../../db')
+                    st.markdown(f"Database {chroma_db_name} has been deleted.")
+        except:
+            pass
+        
+        # Ragatouille
+        st.markdown(show_ragatouille_indexes())
+        try:
+            ragatouille_indexes = [obj.name for obj in show_ragatouille_indexes(format=False)['message']]
+            ragatouille_name = st.selectbox('RAGatouille database to delete', ragatouille_indexes)
+            if delete_buttons:
+                if st.button('Delete RAGatouille database', help='This is permanent!'):
+                    data_processing.delete_index('Ragatouille', ragatouille_name, "Standard", local_db_path='../../db')
+                    st.markdown(f"Index {ragatouille_name} has been deleted.")
+        except:
+            pass
+
 def _format_key_status(key_status:str):
     formatted_status = ""
     for key, value in key_status.items():
